invoice model updated

+++ models/invoice.py
@@ -1,7 +1,7 @@
 import datetime
 
 from pydantic import BaseModel
-from sqlalchemy import Column, Integer, ForeignKey, TIMESTAMP
+from sqlalchemy import Column, Integer, ForeignKey, TIMESTAMP, Double
 
 
 class Invoice(BaseModel):
@@ -9,9 +9,9 @@ class Invoice(BaseModel):
     id = Column(Integer, primary_key=True, index=True)
     user_id = Column(Integer, ForeignKey("user.id"))
     customer_id = Column(Integer, ForeignKey("customer.id"))
-    total = Column(Integer, default=0)
-    discount = Column(Integer, default=0)
-    vat= Column(Integer, default=0)
-    payable = Column(Integer, default=0)
+    total = Column(Double, default=0)
+    discount = Column(Double, default=0)
+    vat= Column(Double, default=0)
+    payable = Column(Double, default=0)
     created_at = Column(TIMESTAMP, nullable=False, default=lambda: datetime.now(datetime.timezone.utc))
     updated_at = Column(TIMESTAMP, nullable=False, default=lambda: datetime.now(datetime.timezone.utc))
